Fix validation logic for oneOf usage.
The original code assumed that the class definitions accessed from
__SCHEMA_TYPES__ would always be scalar; however, both number and
string has the potential to be multiple classes.  As such, restructure
the type checks to build up a list first, then do a validation run.

Additionally, fix the logic so that it actually *runs* the validation
checks if given a oneOf (or anyOf); the previous code would incorrectly
break w/out doing a validation check.