Fix ArrayValidator.validate() caching bug.
The implementation of ArrayValidator would attempt to lazily convert
the validators for items: {'type': 'whatever'} to a list of validators
the same length as the passed in data.

However, if the data changes between validate calls, this invalidly
induces a max/min constraint on the data; thus leave the attribute
as it was, and do the expansion within the function alone (no
side effects).