Fix ref resolution.
If we're in id/schema http://blah/#monkeys , a ref of #/blah should
result in http://blah/#blah ; previous code would resolve it to
http://blah/#monkeys/#blah .

I *suspect* there is a codepent bug- the previous broken resolution
was still resolving in my local usage, despite the fact it shouldn't
have resulted in a usable resolution.  I suspect the issue will be
found if folks dig through id support.

Irregardless, the current logic is incorrect, thus this fix.