Suppress all exceptions from issubclass checks.
In a perfect world, the code would structure it's data structures so
that it never would invoke things like issubclass('string', LiteralValue);
however, it doesn't, and issubclass has an annoying wart where it exposes
any underlying exceptions that occur during the check.

In particular, abc.Meta tries to memoize the results of subclass checks,
using a weakref cache to track it- thus a TypeError can be thrown if
the passed in first argument doesn't support weakrefs (as a list or string
does in this case).

TLDR; introduce a util.safe_issubclass function, use that instead of
issubclass so that any TypeError's encountered are suppressed- end result,
the framework properly reports the jsonp error it's building towards, rather
than puking an indecipherable TypeError from deep within the bowels of abc.Meta.