Property setters; don't store the converted value for primitives.
If we're given a string or list, store the original instance/type-
else we wind up with scenarios where accessing what should be a
simple attribute (say for a string) is instead a ProtocolBase derivative
that isn't directly usable as the original type (say, as a string).

Consuming code shouldn't have to wrap all attribute access in type
conversions; instead, just fix the validation logic to not store
validators of primitives.