bugfix: Support oneOf in array types. (#56)
* bugfix: Support oneOf in array types.

Also make top-level array types expose the same interface as
ProtocolBase (from_json, serialize).

Fixes #49

* bugfix: call validate() when instantiating TypeProxy objects

There are validations that occur based on the presence of properties
(i.e. you cannot assign an object with invalid properties), and there
are different validations that are applied on demand or when serializing
by calling validate() (these are things like 'required' and range
checks). When deciding which type to use from a TypeProxy, we need to
make sure we call validate() because we want to make sure that we
correctly determine which type is valid.