Fix property setting bug when typ is dict
Fixed a bug where a StopIteration exception would be raised because of
improper filtering on the `validators.SCHEMA_TYPE_MAPPING` when
searching for a property's type definition.

Added a regression test for the bug. The test uses a schema where a property's `type` was defined in a `oneOf`.
This hits the code path of the bug.