bugfix: Properly handle attribute deletion. Fixes #165
This resolves a couple of issues related to attribute deletion.

First, all the various del methods now work, including those that use
attribute syntax such as `del foo.bar`.

Second, the handling of attribute deletion works properly for
additionalProperties and standard properties in that for
additionalProperties, the property is just removed, but for standard
properties it is set to None. Standard properties that are `required`
are not permitted to be deleted.