Feature: Fixes #151
This resolves an issue where if a oneOf was hidden behind a $ref
at the toplevel of an schema (I know), then it would fail to load properly.

It turns out this is just because the top level reference was using an
old method for resolving  type proxies. Updating it to use the TypeProxy
object resolved the issue.

Some minor refactoring of names too.