build: Update versioneer from 0.15 to 0.29
One benefit of this update is to support the upcoming Python 3.12
release. Old versioneer versions use configparser.SafeConfigParser.
This was deprecated back in Python 3.2, and being removed in 3.12.

Procedure followed for the update:

- Manually removed the code appended by the old version to the end
  of __init__.py, which the update will not do. (When this is not
  done, it appends improved but similar code after it.)

- In the top-level repository directory (i.e., the directory that
  directly contains setup.py), followed the current instructions
  for a "Vendored mode" installation, by installing the versioneer
  package in a temporary virtual environment and running
  "versioneer install --vendor".

As expected, this reported that the old versioneer.py was replaced,
and that the effect of running the new script was to create
_version.py and append to __init__.py.