workflows: Update black and improve command
This updates black and improves the command used to run it.

- Update repo URL from ambv/black to psf/black in pre-commit hook
- Update black version from 23.3.0 to 23.7.0 in pre-commit hook
- Update black version (same as above) in the linting job
- Make command simpler and more robust, using --extend-exclude

The change to the command is probably the most significant of these
changes. The main reason the command is now more robust is that it
was previously using $() command substitution in a way that would
do the wrong thing on filenames that contain whitespace or special
characters like * (and using grep in a way that would match files
with .py anywhere in their names rather than just at the end).

There are other files than versioneer.py that black should ignore,
because it should ignore files in locations ignored by git, but
black defaults to this behavior (at least in recent versions), and
using --extend-exclude instead of --exclude ensures that everything
black excludes when its exclusion is not customized stays excluded.
This allows the path to be passed just as "." and eliminates the
need to use "git ls-files".