workflows: Update CI action versions
This updates versions of GitHub Actions, *except* in the stale.yml
workflow. The upgraded actions in other workflows are checkout,
setup-python, and gh-action-pypi-publish.

The checkout and setup-python actions were upgraded across major
versions. The advantages outweigh the disadvantages, because:

- The syntax the workflows are using remains supported.
- They were upgraded in some places already, just not all.
- Old Node 12 actions may break around 27 September when GitHub
  Actions removes Node 12 support and runs all actions on Node 16.
- There may be a benefit to bugfixes in the newer versions.

For gh-action-pypi-publish, the minor version upgrade includes
security updates to internal dependencies.

I have not touched stale.yml here, even though the action it uses
is specified to use a very old version, because that workflow is
currently disabled (which appears to have been done automatically
at some point during a lull in activity when no repository activity
was detected for 60 days). Modifying it might re-enable it, which
might not be desired.