cleanup: Unset test_regression_114 executable bit
test/test_regression_114.py, and no other test modules, had its
executable bit set. Since there are no indications that this was
meant to be treated as an executable file on Unix-like systems,
running it successfully as a script would not run the test, and
executing it would typically cause a POSIX-style shell rather than
a Python interpreter to run it (there is no "#!"), that appears
unintentional.

This unsets that executable bit so that git tracks all the test
modules with the same file mode values.