workflows: Avoid duplication in a nicer way
This lists all Python versions in the same list, given as the value
of the python-version key for the matrix, as was done in the past.
However, it still avoids producing two separate experimental and
non-experimental jobs for the same version that is intended to be
experimental, by listing both general and overridden "experimental"
keys in "include", and *not* listing "experimental" outside
"include".

This gives the workflow readability benefit of the old code, while
still fixing the problem that a separate non-experimental job was
also generated.