tests: Fix coverage command so it works on Windows
Although the project and its tests were already fully compatible
with Windows, the coverage command in tox.ini required the pytest
executable to be a Python source code file. On most platforms,
console_scripts stubs are actual Python scripts, but on Windows
they are binaries. "coverage run" doesn't support non-Python files.

This changes the command in tox.ini to use the "coverage run -m"
form instead, which works on all platforms. (This is the approach
shown in the "coverage" documentations for running pytest with it.)