workflows: Fix/update test job names
This adds operating systems to the test job names. Because job
names are being explicitly constructed, this has to be done in
order to give different names to jobs that use the same Python
version but different operating systems.

Because "Build" is really the name of the workflow as a whole,
which defines both a linting job and testing jobs, this also
changes the "build" key for the test jobs to "test", and the
"Build" prefix in the constructed test job names to "Test".