cleanup: Refine/shorten test job names
This shortens the test job names, which became very long due to
recent changes and would often not be fully visible in the GitHub
Actions web-based interface.

The new style is what GitHub Actions would generate automatically,
except that the value of experimental (which is used for
continue-on-error) is included.

This style omits the keys, just showing the values. But if the keys
are also included, the names are often too wide to fully see.