tests: Use importlib.resources for README.md path
This migrates the code in conftest.py that gets a guaranteed usable
path to README.md, from pkg_resources to importlib.resources, which
is its recommended replacement.

The reason to do this is that setuptools, which provides the
pkg_resources module, has deprecated importing pkg_resources.
Although that happened a while ago, the reason to do this now is
that setuptools will no longer be installed by default on Python
3.12. (So the main alternative to migrating to importlib.resources
would be to deepen reliance on a deprecated feature by adding a
setuptools test dependency in tox.ini, just to use the feature.)

This commit does not complete the migration. It preserves the old
approach as well, to assert that both results are the same. This is
temporary, just for the purpose of checking the new code.