tests: Remove the validator.validate mock instead
The preceding commit (f0747ce) shows that mocking and counting
calls to validator.validate verifies the same number of calls to it
as to validator.validate_items (which already had assertions).

However, that validator.validate is called this number of times,
and no more, in the non-strict test, does not appear to be a fact
that the tests in test_regression_143 intend to claim and verify.
There are multiple places where a strictness check can prevent
validate_items from being called. One is in the typed_elems
property getter, which conditionally calls validate, which calls
validate_items, and that's what happens in this test. But validate
calls validate_items conditionally as well. A refactoring of the
code could potentially cause more validate calls to happen in the
non-strict case while still not increasing the call count for
validate_items.

Furthermore, the specific goal of this test is to verify that the
change in #149 (for #143) is and remains effective, and the key
goal there is to avoid unnecessarily revalidating *arrays*, so
validate_items is the specifically relevant function to instrument.