cleanup: Improve order OSes are given in CI matrix
This keeps Ubuntu first but puts macOS before Windows, which I had
intended to do. I believe that order is more common; Ubuntu and
macOS are more similar to each other than they are to Windows; and,
in my experience, for installing and running Python code on CI,
Windows tends to be slower than macOS which is slower than Ubuntu,
so by putting them in order of likely decreasing speed, bugs can
be seen sooner, and the fail-fast behavior will more often cancel
longer-running checks that benefit more from cancellation (because
CI checks tend to start in the order they are listed).