tests: Test on Python 3.12 (locally and on CI)
This adds Python 3.12 tox environments and CI test job.

The 3.11 CI job had been marked experimental. This marks that
non-experimental (like the preceding Python versions' jobs) and
marks the 3.12 CI job experimental. If it fails, the workflow as a
whole will still fail, but the other jobs generated by the CI test
matrix will not be automatically canceled due to a 3.12 failure.

In this commit, setuptools is also added as an explicit test
dependency on 3.12, for the pkg_resources import in conftest.py.
(In 3.12, setuptools, which provides pkg_resources, is no longer
automatically installed in virtual environments.) This facilitates
testing, but it may be better to fix that issue by migrating from
pkg_resources to importlib.resources instead.